Rename courses app to course_metadata
