+++ course_discovery/apps/core/tests/mixins.py
@@ -3,7 +3,7 @@ import logging
 from django.conf import settings
 from elasticsearch import Elasticsearch
 
-from course_discovery.apps.courses.utils import ElasticsearchUtils
+from course_discovery.apps.course_metadata.utils import ElasticsearchUtils
 
 logger = logging.getLogger(__name__)
