+++ course_discovery/apps/course_metadata/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'course_discovery.apps.course_metadata.apps.CourseMetadataConfig'
