+++ course_discovery/apps/course_metadata/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CourseMetadataConfig(AppConfig):
+    name = 'course_metadata'
+    verbose_name = 'Course Metadata'
