+++ course_discovery/apps/course_metadata/config.py
