+++ course_discovery/apps/course_metadata/tests/test_install_es_indexes.py
