--- course_discovery/apps/courses/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'course_discovery.apps.courses.apps.CoursesConfig'
